// src/interfaces/eventObjects/DisplayEventItem.ts
import { BigNumber } from "ethers/utils";

//The generic display item, combining all 4 types of events
export interface DisplayEventItem {
    description: string;   
    date: string;
    timestamp: number;
    avatarSeed: string;
    boldItem0: string;
    description0: string;
    boldItem1: string;
    description1: string
    boldItem2: string;
    description2: string
}